SLIMS-26152: Add an explicit way of providing the certificates to query SLIMS REST through HTTPS.
Old behavior:
When we wanted to query SLIMS REST via HTTPS, the easiest solution was to set the environment variable REQUESTS_CA_BUNDLE="path/to/certificate", but this solution was apparently not used.

New behavior:
In addition to the already existing solution, the constructor of Slims accepts the path to the certificate as argument.